Use the actual ProofOfStore (#1398)
Move types needed outside of "Quorum Store proper" into consensus-types
