+++ consensus/src/quorum_store/batch_reader.rs
@@ -2,17 +2,17 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use crate::network_interface::ConsensusMsg;
-use crate::quorum_store::types::{Data, Batch, PersistedValue, ProofOfStore};
+use crate::quorum_store::types::{Batch, Data, PersistedValue};
 use crate::{
     network::NetworkSender,
     quorum_store::{
         batch_requester::BatchRequester,
-        batch_store::{BatchStoreCommand, LogicalTime},
+        batch_store::BatchStoreCommand,
     },
 };
 use aptos_crypto::HashValue;
-use aptos_types::{validator_signer::ValidatorSigner, PeerId};
-use consensus_types::common::{Round};
+use aptos_types::{PeerId, validator_signer::ValidatorSigner};
+use consensus_types::common::Round;
 use dashmap::DashMap;
 use once_cell::sync::OnceCell;
 use std::collections::HashMap;
@@ -20,13 +20,14 @@ use std::{
     cmp::Reverse,
     collections::BinaryHeap,
     sync::{
+        Arc,
         atomic::{AtomicU64, Ordering},
-        mpsc::{Receiver as SyncReceiver, RecvTimeoutError, SyncSender},
-        Arc, Mutex,
+        mpsc::{Receiver as SyncReceiver, RecvTimeoutError, SyncSender}, Mutex,
     },
     time::Duration,
 };
 use tokio::sync::{mpsc::Sender, oneshot};
+use consensus_types::proof_of_store::{LogicalTime, ProofOfStore};
 
 
 // Make configuration parameter (passed to QuorumStore).
